Remake of the build system. Now using premake5. Added GLFW3, GLEW and RP3D as projects to the workspace. Started incorporating PlatformLayer to the codebase. Rewrote FileReader.

+++ Engone/include/Engone/Utilities/Alloc.h
@@ -6,33 +6,36 @@
 // Not that with ALLOC_DELETE you cannot use a class with a namespace.
 // 'ALLOC_DELETE(rp3d::PhysicsCommon,common)' will give 'common->rp3d::PhysicsCommon' which isn't allowed.
 
-#include <stdint.h>
-#include <cstdlib>
+// #include <cstdlib>
+#include <stdlib.h>
 
 #define ALLOC_NEW(CLASS) new((CLASS*)engone::alloc::malloc(sizeof(CLASS))) CLASS
 #define ALLOC_DELETE(CLASS,VAR) {VAR->~CLASS();engone::alloc::free(VAR,sizeof(CLASS));}
 
 namespace engone {
+	
+	// Todo: MemoryAllocator for game state memory
+	
 	namespace alloc {
-		void* malloc(uint64_t size);
+		void* malloc(uint64 size);
 		// oldSize is used for tracking. you usually keep it somewhere.
-		void* realloc(void* ptr, uint64_t oldSize, uint64_t newSize);
+		void* realloc(void* ptr, uint64 oldSize, uint64 newSize);
 		// size is used for tracking. you usually keep it somewhere.
-		void free(void* ptr, uint64_t size);
-		uint32_t allocatedBytes();
+		void free(void* ptr, uint64 size);
+		uint64 allocatedBytes();
 	}
 	// Does not have a destructor. You need to free the memory with resize(0)
 	template<class T>
 	struct Memory {
-		uint32_t max = 0;
-		uint32_t used = 0; // may be useful to you.
+		uint64 max = 0;
+		uint64 used = 0; // may be useful to you.
 		T* data = nullptr;
 
 		// count is not in bytes.
-		bool resize(uint32_t count) {
+		bool resize(uint64 count) {
 			return ((Memory<char>*)this)->resize(sizeof(T), count);
 		}
 		// function is only defined for Memory<char>
-		bool resize(uint32_t size, uint32_t count);
+		bool resize(uint64 size, uint64 count);
 	};
 }
